requirements.txt
streamlit
langchain
pandas
sqlalchemy
sqlite3
langchain-groq
python-dotenv 
langchain-community   